1.文件命名格式变更 2.添加RingBuffer 3.增加tab的输出设备

+++ ZEngine/Core/Containers/public/Array.h
@@ -10,8 +10,8 @@ class ZArray : public std::vector<DATA_TYPE, ALLOCATOR>
 public:
 
 	//return new elem index if added, else return current num of elem
-	inline int AddUnique(DATA_TYPE elem) {
-		int Num = this->size();
+	inline size_t AddUnique(DATA_TYPE elem) {
+		size_t Num = this->size();
 		auto it = std::find(this->begin(), this->end(), elem);
 		if (it == this->end())
 		{
